Added button for making pw visible.
Sometimes I need that :(.

Moved the dialing directory states to a new struct. The MainWindow struct got overloaded with responsibilites.
This created a bit of a mess - esp. the view method needs the main window and the dialing directory data therefore making it a global method.
(No I don't want it to be added to MainWindow - at least not yet).

At some point the dialing_directory_dialog module needs to be broken down and cleaned up.